requirements.txt
bandit==1.5.1
blessed==1.15.0
botocore==1.12.42
cached-property==1.5.1
cement==2.8.2
certifi==2018.10.15
chardet==3.0.4
Click==7.0
colorama==0.3.9
dnspython==1.15.0
docker==3.5.1
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dodgy==0.1.9
dparse==0.4.1
flake8==3.6.0
Flask==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
idna==2.6
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
MarkupSafe==1.1.0
mccabe==0.6.1
packaging==18.0
pathspec==0.5.5
pbr==5.1.1
pycodestyle==2.4.0
pyflakes==2.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
PyYAML==3.13
requests==2.18.4
safety==1.8.4
semantic-version==2.5.0
six==1.11.0
smmap2==2.0.5
stevedore==1.30.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.54.0
Werkzeug==0.14.1
